Broke some tests a while back because I replaced an instance of a manhattan estimate in A* with the actual distanceTo method of the node, revealing a somewhat systemic problem.
The problem is that algorithms typically use a heuristic, `h`, to guess the distance from a node to the end.  However, this same heuristic is not used to calculate distances between neighboring nodes; a separate `c` cost function is used for that.  I've separated this out in my logic as distanceToGoal (the heuristic) and distanceToNeighbor (the cost function).  They take the same args, and the client code will need to determine whether the neighbor is diagonally adjacent or not (if it cares and is relevant).